chore: 🔧 major refactoring of group-member-list file into individual components
