+++ src/components/groups/group-member-list.tsx
@@ -1,57 +1,28 @@
 "use client";
 
-import { removeGroupMember } from "@actions/group/remove-member/action";
-import { ArrowBack, Check, Close, People, Settings } from "@mui/icons-material";
-import ErrorOutlineIcon from "@mui/icons-material/ErrorOutline";
-import EventAvailableIcon from "@mui/icons-material/EventAvailable";
-import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
-import PersonRemoveIcon from "@mui/icons-material/PersonRemove";
+import { Add, ArrowBack, People, Settings, Share } from "@mui/icons-material";
 import {
 	Avatar,
-	Box,
 	Button,
-	ButtonBase,
 	Dialog,
 	DialogContent,
 	DialogTitle,
-	Drawer,
 	IconButton,
-	Menu,
-	MenuItem,
 	Tab,
 	Tabs,
 	Typography,
-	useMediaQuery,
 } from "@mui/material";
-import { alpha, useTheme } from "@mui/material/styles";
-import {
-	Calendar,
-	Clock,
-	MoreVertical,
-	Plus,
-	Share2,
-	Users,
-} from "lucide-react";
 import Link from "next/link";
-import { useMemo, useState, useTransition } from "react";
+import { useMemo, useState } from "react";
 import { FilterChip } from "@/components/ui/filter-chip";
-import {
-	Select,
-	SelectContent,
-	SelectItem,
-	SelectTrigger,
-	SelectValue,
-} from "@/components/ui/select";
 import { useSnackbar } from "@/components/ui/snackbar-provider";
-import { GroupRole, type SelectGroup } from "@/db/schema";
-import { formatDateToUSNumeric } from "@/lib/availability/utils";
+import type { GroupRole, SelectGroup } from "@/db/schema";
 import { copyTextToClipboard } from "@/lib/clipboard/utils";
-import { isAnchorDateString, WEEKDAYS } from "@/lib/types/chrono";
 import { createGroupInvite } from "@/server/actions/group/invite/create/action";
-import { updateMemberRole } from "@/server/actions/group/update-member-role/action";
-import { nudgePendingMembers } from "@/server/actions/meeting/nudge/action";
 import type { MeetingWithStats } from "@/server/data/groups/queries";
 import { GroupSettingsForm } from "./group-settings-form";
+import { MeetingRow } from "./meeting-row";
+import { MembersList } from "./members-list";
 
 type Member = {
 	userId: string;
@@ -67,611 +38,6 @@ interface GroupMemberListProps {
 	isAdmin: boolean;
 	currentUserId: string;
 	currentMemberId: string;
-}
-
-function DateRange({
-	dates,
-	meetingType,
-}: {
-	dates: string[];
-	meetingType: "dates" | "days";
-}) {
-	if (!dates || dates.length === 0) return <>No dates specified</>;
-
-	if (meetingType === "days") {
-		const weekdayData = dates
-			.map((dateStr) => {
-				const dateString = dateStr.split("T")[0];
-				if (isAnchorDateString(dateString)) {
-					const date = new Date(dateStr);
-					const dayIndex = date.getUTCDay();
-					return { dayIndex, name: WEEKDAYS[dayIndex] };
-				}
-				return null;
-			})
-			.filter(Boolean) as { dayIndex: number; name: string }[];
-
-		if (weekdayData.length > 0) {
-			const sortedWeekdays = weekdayData
-				.sort((a, b) => a.dayIndex - b.dayIndex)
-				.map((day) => day.name);
-			return <>{sortedWeekdays.join(", ")}</>;
-		}
-	}
-
-	if (dates.length === 1) {
-		return <>{formatDateToUSNumeric(new Date(dates[0]))}</>;
-	}
-
-	const sortedDates = [...dates].sort(
-		(a, b) => new Date(a).getTime() - new Date(b).getTime(),
-	);
-
-	const first = formatDateToUSNumeric(new Date(sortedDates[0]));
-	const last = formatDateToUSNumeric(
-		new Date(sortedDates[sortedDates.length - 1]),
-	);
-
-	return <>{`${first} - ${last}`}</>;
-}
-
-function StatusBadge({
-	status,
-}: {
-	status: "actionRequired" | "upcoming" | null;
-}) {
-	const isAction = status === "actionRequired";
-
-	return (
-		<Box
-			sx={(theme) => ({
-				display: "inline-flex",
-				alignItems: "center",
-				gap: 0.5,
-				px: 1.25,
-				py: 0.5,
-				borderRadius: 2,
-				fontSize: "0.75rem",
-				fontWeight: 700,
-				lineHeight: 1,
-				bgcolor: isAction
-					? alpha(theme.palette.error.main, 0.14)
-					: alpha(theme.palette.success.main, 0.14),
-				color: isAction ? "error.main" : "success.main",
-			})}
-		>
-			{isAction ? (
-				<ErrorOutlineIcon sx={{ fontSize: 16, color: "error.main" }} />
-			) : (
-				<EventAvailableIcon sx={{ fontSize: 16, color: "success.main" }} />
-			)}
-			{isAction ? "Action Required" : "Upcoming"}
-		</Box>
-	);
-}
-
-function MeetingRow({
-	meeting,
-	currentMemberId,
-	status,
-}: {
-	meeting: MeetingWithStats;
-	currentMemberId: string;
-	status: "actionRequired" | "upcoming" | null;
-}) {
-	const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-	const [isNudging, setIsNudging] = useState(false);
-	const { showSuccess, showError } = useSnackbar();
-	const open = Boolean(anchorEl);
-	const isOwner = meeting.hostId === currentMemberId;
-	const createdByLabel = isOwner
-		? "Created by You"
-		: `Created by ${meeting.hostName}`;
-
-	return (
-		<div className="relative flex items-center justify-between rounded-xl border-gray-200 border-b px-4 py-4 transition-colors hover:bg-primary/5">
-			<Link
-				href={`/availability/${meeting.id}`}
-				className="absolute inset-0 rounded-xl"
-				aria-label={`Open ${meeting.title}`}
-			/>
-			<div className="pointer-events-none relative z-10 flex flex-1 flex-col gap-1">
-				<div className="flex flex-wrap items-center gap-2">
-					<h3 className="font-medium text-base">{meeting.title}</h3>
-					{status && <StatusBadge status={status} />}
-				</div>
-
-				<div className="flex flex-wrap items-center gap-x-4 text-gray-500 text-sm">
-					<div className="flex items-center gap-1">
-						<Calendar className="size-4" />
-						<span>
-							<DateRange
-								dates={
-									meeting.scheduledDate
-										? [meeting.scheduledDate.toISOString()]
-										: meeting.dates
-								}
-								meetingType={
-									meeting.scheduledDate ? "dates" : meeting.meetingType
-								}
-							/>
-						</span>
-					</div>
-
-					<div className="flex items-center gap-1">
-						<Users className="size-4" />
-						<span>{meeting.totalMembers} members</span>
-					</div>
-
-					<div className="flex items-center gap-1">
-						<Clock className="size-4" />
-						<span>
-							{meeting.respondedCount}/{meeting.totalMembers} responded
-						</span>
-					</div>
-				</div>
-			</div>
-
-			<div className="relative z-20 ml-4 flex items-center gap-3">
-				<p className="pointer-events-none whitespace-nowrap font-medium text-gray-400 text-xs uppercase tracking-wide">
-					{createdByLabel}
-				</p>
-				{isOwner && (
-					<>
-						<IconButton
-							onClick={(e) => {
-								e.stopPropagation();
-								setAnchorEl(open ? null : e.currentTarget);
-							}}
-						>
-							<MoreVertical className="size-5 text-gray-500" />
-						</IconButton>
-
-						<Menu
-							anchorEl={anchorEl}
-							open={open}
-							onClose={() => setAnchorEl(null)}
-							transformOrigin={{ horizontal: "right", vertical: "top" }}
-							anchorOrigin={{ horizontal: "right", vertical: "bottom" }}
-						>
-							<MenuItem
-								disabled={isNudging}
-								onClick={async (e) => {
-									e.stopPropagation();
-									setAnchorEl(null);
-									setIsNudging(true);
-									try {
-										const result = await nudgePendingMembers(meeting.id);
-										if (result.success) {
-											showSuccess(result.message);
-										} else {
-											showError(result.message);
-										}
-									} catch {
-										showError("Failed to send nudge. Please try again.");
-									} finally {
-										setIsNudging(false);
-									}
-								}}
-							>
-								<People className="mr-2 size-4" />
-								Nudge Members
-							</MenuItem>
-						</Menu>
-					</>
-				)}
-			</div>
-		</div>
-	);
-}
-
-function MemberAvatar({ email }: { email: string }) {
-	const initials = email.split("@")[0]?.slice(0, 2).toUpperCase() ?? "??";
-	return (
-		<div className="flex size-12 shrink-0 items-center justify-center rounded-full bg-slate-200 font-medium text-slate-600 text-sm">
-			{initials}
-		</div>
-	);
-}
-
-function AdminMemberRow({
-	member,
-	groupId,
-	isSelf,
-}: {
-	member: Member;
-	groupId: string;
-	isSelf: boolean;
-}) {
-	const [isPending, startTransition] = useTransition();
-	const [showRemoveDialog, setShowRemoveDialog] = useState(false);
-	const [showPermissionsSheet, setShowPermissionsSheet] = useState(false);
-
-	const theme = useTheme();
-	const isMobile = useMediaQuery(theme.breakpoints.down("md"));
-
-	const { showSuccess, showError } = useSnackbar();
-
-	function handleRoleChange(value: string) {
-		startTransition(async () => {
-			const result = await updateMemberRole({
-				groupId,
-				targetUserId: member.userId,
-				role: value as GroupRole,
-			});
-
-			if (result?.success) {
-				showSuccess(result.message);
-			} else if (result?.message) {
-				showError(result.message);
-			}
-		});
-	}
-
-	async function handleRemoveMember() {
-		const result = await removeGroupMember(groupId, member.userId);
-
-		if (result.success) {
-			showSuccess(result.message);
-			setShowRemoveDialog(false);
-		} else {
-			showError(result.message);
-		}
-	}
-
-	return (
-		<>
-			<div
-				style={{
-					borderBottom: `1px solid ${theme.palette.divider}`,
-				}}
-				className="flex h-[83px] items-center justify-between px-4 transition-colors hover:bg-gray-50/50"
-			>
-				<div className="flex items-center gap-4">
-					<MemberAvatar email={member.email} />
-					<span
-						style={{
-							color: theme.palette.text.primary,
-						}}
-						className="font-medium text-base"
-					>
-						{member.email.split("@")[0]}
-
-						{isSelf && (
-							<span
-								style={{
-									color: theme.palette.text.secondary,
-								}}
-								className="ml-2 text-xs"
-							>
-								(you)
-							</span>
-						)}
-					</span>
-				</div>
-
-				{isMobile ? (
-					<Button
-						disabled={isPending}
-						onClick={() => setShowPermissionsSheet(true)}
-						disableRipple
-						sx={{
-							textTransform: "none",
-							minWidth: "auto",
-							padding: 0,
-							color: "text.primary",
-							fontSize: "0.875rem",
-							fontWeight: 500,
-							gap: 0.5,
-							"&.Mui-disabled": {
-								opacity: 0.5,
-							},
-						}}
-						endIcon={
-							<KeyboardArrowDownIcon
-								sx={{
-									fontSize: 16,
-									color: "text.secondary",
-								}}
-							/>
-						}
-					>
-						{member.role === GroupRole.ADMIN ? "Admin" : "Member"}
-					</Button>
-				) : (
-					<div className="flex items-center gap-2">
-						<Select
-							value={member.role ?? GroupRole.MEMBER}
-							onValueChange={handleRoleChange}
-							disabled={isSelf || isPending}
-						>
-							<SelectTrigger
-								style={{
-									backgroundColor: theme.palette.background.default,
-									color: theme.palette.text.primary,
-									border: `1px solid ${theme.palette.divider}`,
-								}}
-								className="h-9 w-[140px] rounded-lg text-sm"
-							>
-								<SelectValue />
-							</SelectTrigger>
-
-							<SelectContent>
-								<SelectItem value={GroupRole.ADMIN}>Admin</SelectItem>
-								<SelectItem value={GroupRole.MEMBER}>Member</SelectItem>
-							</SelectContent>
-						</Select>
-
-						{!isSelf && (
-							<IconButton
-								color="error"
-								onClick={() => setShowRemoveDialog(true)}
-							>
-								<PersonRemoveIcon />
-							</IconButton>
-						)}
-					</div>
-				)}
-			</div>
-
-			{/* Mobile member permissions settings */}
-			<Drawer
-				anchor="bottom"
-				open={showPermissionsSheet}
-				onClose={() => setShowPermissionsSheet(false)}
-				slotProps={{
-					paper: {
-						sx: {
-							borderTopLeftRadius: 20,
-							borderTopRightRadius: 20,
-							p: 3,
-						},
-					},
-				}}
-			>
-				<div className="flex flex-col gap-6">
-					<div className="flex items-center justify-between">
-						<Typography variant="h6" fontWeight={700}>
-							Permission Settings
-						</Typography>
-
-						<IconButton onClick={() => setShowPermissionsSheet(false)}>
-							<Close />
-						</IconButton>
-					</div>
-
-					<div className="flex items-center gap-4">
-						<MemberAvatar email={member.email} />
-
-						<div className="flex flex-col">
-							<Typography fontWeight={600}>
-								{member.email.split("@")[0]}
-							</Typography>
-
-							<Typography variant="body2" color="text.secondary">
-								{member.email}
-							</Typography>
-						</div>
-					</div>
-
-					<div className="flex flex-col">
-						{[
-							{
-								label: "Member",
-								description: "Can view and comment on content",
-								value: GroupRole.MEMBER,
-							},
-							{
-								label: "Admin",
-								description: "Full access to all features",
-								value: GroupRole.ADMIN,
-							},
-						].map((role) => {
-							const isSelected = member.role === role.value;
-
-							return (
-								<ButtonBase
-									key={role.value}
-									onClick={() => {
-										handleRoleChange(role.value);
-										setShowPermissionsSheet(false);
-									}}
-									disableRipple
-									sx={{
-										width: "100%",
-										display: "flex",
-										alignItems: "flex-start",
-										justifyContent: "space-between",
-										textAlign: "left",
-										borderRadius: 2,
-										p: 1.5,
-										"&:hover": {
-											backgroundColor: "transparent",
-										},
-									}}
-								>
-									<div>
-										<Typography fontWeight={600}>{role.label}</Typography>
-
-										<Typography
-											variant="body2"
-											color="text.secondary"
-											sx={{ mt: 0.25 }}
-										>
-											{role.description}
-										</Typography>
-									</div>
-
-									{isSelected && (
-										<Check
-											sx={{
-												color: "primary.main",
-												fontSize: 20,
-												mt: 0.25,
-											}}
-										/>
-									)}
-								</ButtonBase>
-							);
-						})}
-					</div>
-
-					{/* Remove member */}
-					{!isSelf && (
-						<ButtonBase
-							onClick={() => {
-								setShowPermissionsSheet(false);
-								setShowRemoveDialog(true);
-							}}
-							sx={{
-								display: "flex",
-								flexDirection: "column",
-								alignItems: "flex-start",
-								justifyContent: "flex-start",
-								textAlign: "left",
-								width: "100%",
-								borderRadius: 2,
-								p: 1.5,
-								gap: 0.5,
-							}}
-						>
-							<Typography color="error" fontWeight={600}>
-								Remove Account
-							</Typography>
-
-							<Typography
-								variant="body2"
-								color="text.secondary"
-								sx={{ maxWidth: "100%" }}
-							>
-								This account will be permanently removed from the group & no
-								longer have access to meetings.
-							</Typography>
-						</ButtonBase>
-					)}
-				</div>
-			</Drawer>
-
-			{/* Remove confirmation */}
-			<Dialog
-				open={showRemoveDialog}
-				onClose={() => setShowRemoveDialog(false)}
-				maxWidth="xs"
-				fullWidth
-			>
-				<DialogTitle
-					sx={{
-						display: "flex",
-						alignItems: "center",
-						justifyContent: "space-between",
-						fontWeight: 700,
-					}}
-				>
-					Remove Account
-					<IconButton onClick={() => setShowRemoveDialog(false)}>
-						<Close />
-					</IconButton>
-				</DialogTitle>
-
-				<DialogContent>
-					<div className="flex flex-col gap-6 pt-2">
-						<Typography variant="body2" color="text.secondary">
-							This account will be permanently removed from the group & no
-							longer have access to meetings.
-						</Typography>
-
-						<div className="flex items-center gap-4">
-							<MemberAvatar email={member.email} />
-
-							<div className="flex flex-col">
-								<Typography fontWeight={600}>
-									{member.email.split("@")[0]}
-								</Typography>
-
-								<Typography variant="body2" color="text.secondary">
-									{member.email}
-								</Typography>
-							</div>
-						</div>
-
-						<Button
-							fullWidth
-							color="error"
-							variant="contained"
-							onClick={handleRemoveMember}
-							disabled={isPending}
-						>
-							Remove Account
-						</Button>
-					</div>
-				</DialogContent>
-			</Dialog>
-		</>
-	);
-}
-
-function MembersList({
-	members,
-	isAdmin,
-	groupId,
-	currentUserId,
-}: {
-	members: Member[];
-	isAdmin: boolean;
-	groupId: string;
-	currentUserId: string;
-}) {
-	const theme = useTheme();
-	return (
-		<div className="mt-4">
-			<p className="mb-2 px-4 font-bold text-[#969696] text-xs uppercase tracking-wide">
-				All Members ({members.length})
-			</p>
-
-			<div className="flex flex-col">
-				{members.map((member) =>
-					isAdmin ? (
-						<AdminMemberRow
-							key={member.userId}
-							member={member}
-							groupId={groupId}
-							isSelf={member.userId === currentUserId}
-						/>
-					) : (
-						<div
-							key={member.userId}
-							style={{ borderBottom: `1px solid ${theme.palette.divider}` }}
-							className="flex h-[83px] items-center justify-between px-4 transition-colors hover:bg-gray-50/50"
-						>
-							<div className="flex items-center gap-4">
-								<MemberAvatar email={member.email} />
-								<span
-									style={{ color: theme.palette.text.primary }}
-									className="font-medium text-base"
-								>
-									{member.email.split("@")[0]}
-									{member.userId === currentUserId && (
-										<span
-											style={{ color: theme.palette.text.secondary }}
-											className="ml-2 text-xs"
-										>
-											(you)
-										</span>
-									)}
-								</span>
-							</div>
-							<span
-								style={{ color: theme.palette.text.secondary }}
-								className="font-medium text-sm italic"
-							>
-								{member.role === GroupRole.ADMIN ? "Admin" : "Member"}
-							</span>
-						</div>
-					),
-				)}
-			</div>
-		</div>
-	);
 }
 
 export function GroupMemberList({
@@ -892,7 +258,7 @@ export function GroupMemberList({
 									},
 								}}
 							>
-								<Plus className="size-5" />
+								<Add className="size-5" />
 							</Button>
 						</Link>
 						<Button
@@ -913,7 +279,7 @@ export function GroupMemberList({
 								},
 							}}
 						>
-							<Share2 className="size-5" />
+							<Share className="size-5" />
 						</Button>
 					</div>
 				</div>
@@ -936,7 +302,7 @@ export function GroupMemberList({
 						>
 							<Button
 								variant="contained"
-								startIcon={<Plus className="size-4" />}
+								startIcon={<Add className="size-4" />}
 							>
 								Create New Meeting
 							</Button>
@@ -945,7 +311,7 @@ export function GroupMemberList({
 						{/* Desktop Share */}
 						<Button
 							variant="outlined"
-							startIcon={<Share2 className="size-4" />}
+							startIcon={<Share className="size-4" />}
 							onClick={handleCreateInviteLink}
 							disabled={!canShareInvites || isCreatingInvite}
 							sx={{
